fix minor bug in ui
